+++ CHANGELOG.rst
@@ -1,6 +1,17 @@
-Version 0.9.0 - Development
+
+Version 1.0.0 - Development
 =======================================
-* Introduced a new Ajax decorator called ``rearrange`` which adds drag & drop functionalities during page editing. 
+* **BACKWARD INCOMPATIBLE CHANGES**
+* This is the first release not dependent on django.
+* Object Relational mapping can be done with ``django`` or ``python-stdnet`` or none of them.
+* Other ORMs can be registered to the library::
+
+    from djpcms.core.orms register_wrapper
+    
+    register_wrapper('path.to.my.ormwrapper')
+    
+* Introduced a new Ajax decorator called ``rearrange`` which adds drag & drop
+  functionalities during page editing.
 * Added abstraction for ``User``, ``Request`` and ``Response`` classes.
 * Added ``save as new`` functionality to :class:`djpcms.views.appview.EditView`.
 * Added `rst` markup handled by sphinx_ in :mod:`djpcms.contrib.flowrepo.markups` module.